Improvements to the API and implementation, added document generation.
Updated the implementation and header files after code review.
Implementation does not convert from int to float.  Default animations
are disabled while laying out the grid.  Removed retain cycle from
example view; this delegate was not used but best not to leave that in
the project.  General source formatting for readability.  Doxygen: Added
documentation generation as a build target, assumes a default build and
install of doxygen.  Formatted comments can now be added to header
files.