Enable "modules" support and better warnings
This update to the project configuration adds the more aggressive error
and warning settings. This should be similar to the default settings
for a new project in Xcode. Enables some new and old compiler features
to get more help from the compiler. Modules support allows for faster
build times and avoids user defined header conflicts.